Phase 3 validation: updates for msonet yamls (#250)
Finished Phase 3 testing with great results. Since such extensive
testing is done during Phase 2, there is little if any that needs to be
done for Phase 3. All of these updates are related to using the updated
GSI case (which more accurately reflects the RRFSv1 configuration) or
were formatting or improved comments. Phase 3 results start in the
associated issue with this
https://github.com/NOAA-EMC/RDASApp/issues/80#issuecomment-2523984973

For future: It would be nice to be able to read the info from convinfo
instead of hard coding them into each yaml. I bet that capability is
part of JCB.

 Issue(s) addressed #232