Phase 3 validation: sonde (adpupa 120/220) (#256)
Finished Phase 3 testing with great results. Since such extensive
testing is done during Phase 2, there is little if any that needs to be
done for Phase 3. All of these updates are related to using the updated
GSI case (which more accurately reflects the RRFSv1 configuration).
Phase 3 results start in the associated issue with this
https://github.com/NOAA-EMC/RDASApp/issues/232#issuecomment-2552253479.

## Issue(s) addressed:  https://github.com/NOAA-EMC/RDASApp/issues/232